+++ protected/whitelist.php
@@ -69,7 +69,8 @@ function print_whitelist() {
     if ($db_result[RESPONSE_ARGUMENT::CODE] == RESPONSE_ERROR::NONE) {
         $institutions = $db_result[RESPONSE_ARGUMENT::VALUE];
         if (count($institutions) > 0) {
-            print "<p>(Check things if you want to delete them)</p>";
+	    print "<p>Listed institutions are whitelisted; users using a confirmed email address from this insitution may use GENI without manual approval.</p>";
+            print "<p>Check a domain to select it for deletion, or add a new institution below.</p>";
             print "<form action='whitelist.php' method='POST'><ul style='list-style-type:none'>";
             $i = 0;
             foreach ($institutions as $inst) {
@@ -89,18 +90,9 @@ function print_whitelist() {
     }
 }
 
+require_once("header.php");
+show_header("GENI IdP Whitelisted Domain Management", array());
 ?>
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="utf-8">
-<title>GENI IDP Whitelist</title>
-<link href="https://fonts.googleapis.com/css?family=Open+Sans:400,700" rel="stylesheet" type="text/css">
-<link type="text/css" href="geni-ar.css" rel="Stylesheet"/>
-<script type='text/javascript' charset='utf8' src='https://ajax.googleapis.com/ajax/libs/jquery/2.1.4/jquery.min.js'></script>
-
-</head>
 
 <script type="text/javascript">
     $(document).ready(function(){
@@ -113,10 +105,9 @@ function print_whitelist() {
         });
     });
 </script>
-<body>
 
+<h2  style='margin-top: 80px;' class='card'>Whitelisted Domains</h2>
 <div id="content" class='card' style="width:500px; margin: 30px auto">
-<h2>Whitelist page</h2>
 
 <?php